chore: refactor, fmt, clippy run
